move socket and getsockopt into their own files
